Initial Commit of TrackFriends App
iOS app which tracks the multiple friends using location based services
and send messages which will be sent to all of them. This is
particularly useful when group of friends want to get together and know
each other updates using this app ..

This is initial idea..App can also have show near by interesting to
places to go and hangout. etc…